requirements.txt
asgiref==3.2.10
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
pep8==1.7.1
phonenumbers==8.12.11
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.2.0
